requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
kafka-python
slack-sdk
requests
scikit-learn
pandas
salesforce-merlion
pymongo